+++ .idea/inspectionProfiles/Project_Default.xml
@@ -4,7 +4,11 @@
     <inspection_tool class="DuplicatedCode" enabled="true" level="WEAK WARNING" enabled_by_default="true">
       <Languages>
         <language minSize="47" name="Rust" />
+        <language minSize="204" name="TypeScript" />
       </Languages>
+    </inspection_tool>
+    <inspection_tool class="JSDuplicatedDeclaration" enabled="true" level="INFORMATION" enabled_by_default="true" editorAttributes="INFORMATION_ATTRIBUTES">
+      <scope name="Open Files" level="WARNING" enabled="true" />
     </inspection_tool>
   </profile>
 </component>
