+++ pax-chassis-web/pax-dev-harness-web/src/classes/messages/image-load-patch.ts
@@ -1,0 +1,14 @@
+export class ImageLoadPatch {
+    public id_chain?: number[];
+    public path?: string;
+
+    fromPatch(jsonMessage: any) {
+        this.id_chain = jsonMessage["id_chain"];
+        this.path = jsonMessage["path"];
+    }
+
+    cleanUp(){
+        this.id_chain = [];
+        this.path = '';
+    }
+}
