+++ pax-chassis-web/pax-dev-harness-web/src/types/stats-js.d.ts
@@ -1,0 +1,1 @@
+declare module 'src/types/stats-js';
