+++ pax-chassis-web/pax-dev-harness-web/src/utils/constants.ts
@@ -1,0 +1,8 @@
+export const MOUNT_ID = "mount";
+export const NATIVE_OVERLAY_CLASS = "native-overlay";
+export const CANVAS_CLASS = "canvas";
+export const SCROLLER_CONTAINER = "scroller-container"
+export const INNER_PANE = "inner-pane"
+export const NATIVE_LEAF_CLASS = "native-leaf";
+export const NATIVE_CLIPPING_CLASS = "native-clipping";
+export const CLIP_PREFIX = "clip"
