--- pax-chassis-web/pax-dev-harness-web/stats-js.d.ts
@@ -1,1 +1,0 @@
-declare module 'stats-js';
