+++ pax-chassis-web/pax-dev-harness-web/tsconfig.json
@@ -6,5 +6,6 @@
     "lib": ["dom", "es2015"],
     "strict": true,
     "esModuleInterop": true,
+    "sourceMap": true,
   }
 }
