+++ pax-chassis-web/src/lib.rs
@@ -10,6 +10,7 @@ use std::collections::HashMap;
 use piet::TextStorage;
 
 use piet_web::WebRenderContext;
+use serde::de::Unexpected::Option;
 
 use pax_core::{InstanceRegistry, PaxEngine};
 
@@ -60,9 +61,6 @@ pub struct PaxChassisWeb {
 #[wasm_bindgen]
 impl PaxChassisWeb {
 
-
-
-
     //called from JS, this is essentially `main`
     pub fn new() -> Self {
 
@@ -86,8 +84,8 @@ impl PaxChassisWeb {
         }
     }
 
-    pub fn add_context(&mut self, scroller_id: Option<Vec<u64>>, z_index: u32) {
-        let id = PaxChassisWeb::generate_location_id(scroller_id, z_index);
+    pub fn add_context(&mut self, id: String) {
+        pax_runtime_api::log(format!("added context {}", id).as_str());
         let window = window().unwrap();
         let dpr = window.device_pixel_ratio();
         let document = window.document().unwrap();
@@ -114,9 +112,8 @@ impl PaxChassisWeb {
     pub fn sendViewportUpdate(&mut self, width: f64, height: f64){
         self.engine.borrow_mut().set_viewport_size((width, height));
     }
-
-    pub fn remove_context(&mut self, scroller_id: Option<Vec<u64>>, z_index: u32) {
-        let id = PaxChassisWeb::generate_location_id(scroller_id, z_index);
+    pub fn remove_context(&mut self, id: String) {
+        pax_runtime_api::log(format!("removed context {}", id).as_str());
         self.drawing_contexts.remove(&id);
     }
 
@@ -370,11 +367,6 @@ impl PaxChassisWeb {
             len: bytes.len(),
         }
     }
-
-    pub fn generate_location_id(scroller_id: Option<Vec<u64>>, z_index: u32) -> String {
-        ZIndex::generate_location_id(scroller_id, z_index)
-    }
-
 }
